# More .vcxproj cleanup / NP21 core performance flags