--- a/np21_i386c/ia32/cpu.h
+++ b/np21_i386c/ia32/cpu.h
@@ -129,6 +129,7 @@
 #define USE_SSE2
 #define USE_SSE3
 #define USE_TSC
+#define USE_PAGING
 #define USE_FASTPAGING
 #define USE_VME
 #define IA32_REBOOT_ON_PANIC
@@ -1088,7 +1089,11 @@ void CPUCALL set_eflags(UINT32 new_flags, UINT32 mask);
 #define	CPU_STAT_SS32		CPU_STATSAVE.cpu_stat.ss_32
 #define	CPU_STAT_RESETREQ	CPU_STATSAVE.cpu_stat.resetreq
 #define	CPU_STAT_PM		CPU_STATSAVE.cpu_stat.protected_mode
-#define	CPU_STAT_PAGING		CPU_STATSAVE.cpu_stat.paging
+#ifdef USE_PAGING
+	#define	CPU_STAT_PAGING	CPU_STATSAVE.cpu_stat.paging
+#else
+	#define CPU_STAT_PAGING	false
+#endif
 #define	CPU_STAT_VM86		CPU_STATSAVE.cpu_stat.vm86
 #define	CPU_STAT_WP		CPU_STATSAVE.cpu_stat.page_wp
 #define	CPU_STAT_CPL		CPU_CS_DESC.rpl