--- a/np21_i386c/ia32/ia32.mcr
+++ b/np21_i386c/ia32/ia32.mcr
@@ -73,14 +73,16 @@ do { \
 /*
  * clock
  */
-#define	CPU_WORKCLOCK(clock) \
-do { \
-	CPU_REMCLOCK -= (clock); \
-} while (/*CONSTCOND*/ 0)
 
 #define	CPU_HALT() \
 do { \
 	CPU_REMCLOCK = -1; \
+} while (/*CONSTCOND*/ 0)
+
+#ifdef USE_CLOCK
+#define	CPU_WORKCLOCK(clock) \
+do { \
+	CPU_REMCLOCK -= (clock); \
 } while (/*CONSTCOND*/ 0)
 
 #define	IRQCHECKTERM() \
@@ -90,6 +92,10 @@ do { \
 		CPU_REMCLOCK = 0; \
 	} \
 } while (/*CONSTCOND*/ 0)
+#else
+#define	CPU_WORKCLOCK(clock)
+#define	IRQCHECKTERM()
+#endif
 
 
 /*