… msdos.vcxproj