These are the remaining changes I added to the [ReC98](https://github.com/nmlgc/ReC98)-specific fork of MS-DOS Player. They might not necessarily fit this fork, so take whatever looks good to you.

* The first three commits remove all redundancy from the .vcxproj files, so that 38ad39cf9a8789a39b2b7052b11ba24ab27f798b and f7b94d9fb8c3586170055da655a7546a6fa5fecc can apply their settings globally by just modifying a single line.
* 4856647360cb136b86c249cae2a613f13b737617 is a bugfix that allows all `USE_` flags to be `#undef`'d without compile errors.
* aa364cf98a3e1bc026ca343564c535058a578f9b and 490a12bc467c985e636f6262abd7a9ca7f305836 add additional flags for disabling paging (for programs that use Protected Mode segmentation instead) or cycle counting (for programs that don't need to communicate with emulated hardware) in the NP21 core. Some performance numbers from a full rebuild of the ReC98 codebase:

   |                        | Generic build | PGO build |
   | ---------------------: | ------------: | --------: |
   | MAME core              | 50.854s       | 34.885s   |
   | NP21 core, current `vt` branch    | 34.398s       | 29.184s   |
   | + `#undef` all current | 34.276s       | 27.833s   |
   | + `#undef USE_PAGING`  | 31.646s       | 26.356s   |
   | + `#undef USE_CLOCK`   | 26.691s       | 22.599s   |